travaia-ai/backend/interview-session-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
slowapi==0.1.9
redis==5.0.1
google-cloud-firestore==2.13.1
google-cloud-pubsub==2.18.4
google-cloud-storage==2.10.0
google-cloud-bigquery==3.13.0
firebase-admin==6.2.0
structlog==23.2.0
httpx==0.25.2
python-dotenv==1.0.0
tenacity==8.2.3
python-socketio==5.10.0
aiofiles==23.2.1
requests==2.31.0
livekit==0.10.0
livekit-api==1.0.5
PyJWT==2.8.0
websockets==12.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
celery==5.3.4
prometheus-client==0.19.0